+++ Etomo/src/etomo/type/MetaData.java
@@ -29,6 +29,12 @@ import etomo.util.Utilities;
  * @version $Revision$
  *
  * <p> $Log$
+ * <p> Revision 3.68  2011/02/28 22:45:31  sueh
+ * <p> bug# 1452 Making imageRotation double.
+ * <p>
+ * <p> Revision 3.61.2.2  2011/02/25 00:18:13  sueh
+ * <p> bug# 1452 Merged changes from head into 4.1.
+ * <p>
  * <p> Revision 3.67  2011/02/24 23:36:53  sueh
  * <p> bug# 1452 imageRotation needs to be double everywhere.
  * <p>
@@ -408,8 +414,8 @@ public final class MetaData extends BaseMetaData implements ConstMetaData {
   private boolean useLocalAlignmentsA = true;
   private boolean useLocalAlignmentsB = true;
   private double fiducialDiameter = Double.NaN;
-  private double imageRotationA = Float.NaN;
-  private double imageRotationB = Float.NaN;
+  private double imageRotationA = Double.NaN;
+  private double imageRotationB = Double.NaN;
   private int binning = 1;
 
   private boolean fiducialessAlignmentA = false;
@@ -1005,7 +1011,7 @@ public final class MetaData extends BaseMetaData implements ConstMetaData {
     this.fiducialDiameter = Double.parseDouble(fiducialDiameter);
   }
 
-  public void setImageRotation(float rotation, AxisID axisID) {
+  public void setImageRotation(double rotation, AxisID axisID) {
     if (axisID == AxisID.SECOND) {
       imageRotationB = rotation;
     }
@@ -1016,10 +1022,10 @@ public final class MetaData extends BaseMetaData implements ConstMetaData {
 
   public void setImageRotation(String rotation, AxisID axisID) {
     if (axisID == AxisID.SECOND) {
-      imageRotationB = Float.parseFloat(rotation);
+      imageRotationB = Double.parseDouble(rotation);
     }
     else {
-      imageRotationA = Float.parseFloat(rotation);
+      imageRotationA = Double.parseDouble(rotation);
     }
   }
 
@@ -1307,9 +1313,9 @@ public final class MetaData extends BaseMetaData implements ConstMetaData {
     // Read in the old single image rotation or the newer separate image
     // rotation for each axis
     String strOldRotation = props.getProperty(group + "ImageRotation", "0.0");
-    imageRotationA = Float.parseFloat(props.getProperty(group
+    imageRotationA = Double.parseDouble(props.getProperty(group
         + "ImageRotationA", strOldRotation));
-    imageRotationB = Float.parseFloat(props.getProperty(group
+    imageRotationB = Double.parseDouble(props.getProperty(group
         + "ImageRotationB", strOldRotation));
     excludeProjectionsA = props.getProperty(group + "AxisA.ExcludeProjections",
         "");
